Drop generated version.hpp
While on one hand this is (presumably) "addressing" upstream's deficiency
(the lack of the version macro), on the other you are making the package
no longer match what is available via the system package manager e.g.,
on Debian/Ubuntu, etc. And in build2 you are normally expected to be
able to satisfy any dependency with a system-installed package (for
details, see:
https://build2.org/build2-toolchain/doc/build2-toolchain-intro.xhtml#guide-system-deps)

So the better way to address this is to ask upstream to add a suitable
macro.